Filter results of FolderFtsDocset on being an accessible document
This tests each fts result on
- having a nodetype that is not a container
- being publicly accessible on current day

These testings slows down the performance of the query.
It's not a (RUM-)index-only-scan anymore.

Example timing data: 1100ms instead of 300ms

In the future we may want to mitigate this slowdown
by trying to return to an index-only-scan.
Maybe this is possible by putting (and maintaining) additional
informations into the index.