Drop duplicate fts results
We have now combined indexes for attrs and fulltext.
They are currently designed as functional indexes
(aka as indexes on expressions).
This has the effect that each document is referenced
twice in the indexes.

Therefore we need to eliminate possible duplicates in the result sets.

The elimination is implemented in function aux.fts_documents_distinct
as an extra step in order not to spoil the pipelined query evaluation plan.

Please note:
In the future we may change the design of the index,
such that this elimination is no longer necessary.