Use websearch_to_tsquery if available
We want to offer "websearch" syntax for FTS queries, like `signal -"segmentation fault"`.

PostgreSQL version 11 introduces the function websearch_to_tsquery for this purpose, which we now use if available.

We check for availability of this function because (for the time being) we want to be backward compatible with PostgreSQL version 10.