Use union types for wrapping Id types
Formerly we used tuples with some contrived structure to make it
somewhat unique, because tuples are comparable by default and can
be used as keys of core's Dict type.

Now we switched to using Sort.Dict instead of Dict, so the keys don't
need to comparable anymore. Instead we use our own ordering
functions in combination with Sort.Dict.