Switch to Elm version 0.19.1 to avoid a compiler bug
The Elm 0.19.0 compiler crashes with the current source code
when used with the options "--debug --output ...".
This is a known bug in the Elm compiler that is fixed in version 0.19.1.

As of today this version isn't published on npm.
For now we use the binary `elm-0.19.1-beta-1`,
manually linked as `frontend/node_modules/.bin/`