Opaque wrapper for model fields that are used for debugging only
We want to be able to display the current Needs of the app in the Elm Debugger.
So we store the Needs in App.Model for this reason only.

By wrapping the type within the new type DebugInfo we make clear that this field cannot be referenced anywhere else in the code.