Expose all Context types
These types are always type aliases, i.e. records.
So it is not strictly necessary to expose them if the parent module
doesn't use it in a type annotation.

But for documentation reasons (via doc-preview) it's nicer to have the
Context types included in the module's doc page.

So we always should expose these type aliases.