Avoid repeated queries on sorting-independent docsets
FTS queries can be sorted either by rank or by date.
The sorting should only affect the document listing.
Sorting is irrelevant for queries on docsets like folder counts or facets.
These queries shouldn't be repeated when the sorting is changed.

For this we make the relevant cache tables ignore the FtsSorting parameter.